+++ README.md
@@ -1,40 +1,70 @@
-This is my portfolio made in React.
+# Getting Started with Create React App
 
-Libraries used: 
- - React
- - material-ui/core
- 
- This took me around 20 hours to build. I had experience with React, but none with material-ui. It was great figuring how to use these tools, and how to manipulate css as well. 
+This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-<!---
-// https://giant.gfycat.com/SaneDisguisedChital.webm
-// https://giant.gfycat.com/SaneDisguisedChital.mp4
-// Unity zombie game
+## Available Scripts
 
-// https://giant.gfycat.com/ScrawnyGrandioseHeterodontosaurus.webm
-// https://giant.gfycat.com/ScrawnyGrandioseHeterodontosaurus.mp4
-// Unity Shattered Iron
+In the project directory, you can run:
 
-// https://giant.gfycat.com/HeftyScalyInvisiblerail.webm
-// https://giant.gfycat.com/HeftyScalyInvisiblerail.mp4
-// Hairspray
+### `npm start`
 
+Runs the app in the development mode.\
+Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
 
-// https://giant.gfycat.com/SphericalTanHerculesbeetle.webm
-// https://giant.gfycat.com/SphericalTanHerculesbeetle.mp4
-// GiveMeMyMoney
+The page will reload if you make edits.\
+You will also see any lint errors in the console.
 
+### `npm test`
 
-// https://giant.gfycat.com/JubilantImpressiveDanishswedishfarmdog.webm
-// https://giant.gfycat.com/JubilantImpressiveDanishswedishfarmdog.mp4
-// LambdaHairSchool
+Launches the test runner in the interactive watch mode.\
+See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
 
+### `npm run build`
 
-// https://giant.gfycat.com/FeistyDescriptiveEyas.webm
-// https://giant.gfycat.com/FeistyDescriptiveEyas.mp4
-// Tetris
+Builds the app for production to the `build` folder.\
+It correctly bundles React in production mode and optimizes the build for the best performance.
 
-// https://giant.gfycat.com/.webm
-// https://giant.gfycat.com/.mp4
-// example
---->
+The build is minified and the filenames include the hashes.\
+Your app is ready to be deployed!
+
+See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+
+### `npm run eject`
+
+**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
+
+If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
+
+Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
+
+You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
+
+## Learn More
+
+You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
+
+To learn React, check out the [React documentation](https://reactjs.org/).
+
+### Code Splitting
+
+This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
+
+### Analyzing the Bundle Size
+
+This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
+
+### Making a Progressive Web App
+
+This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
+
+### Advanced Configuration
+
+This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
+
+### Deployment
+
+This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
+
+### `npm run build` fails to minify
+
+This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
