--- TODO.md
@@ -1,47 +1,0 @@
-# Overall
-- [x] Create Carosel/slideshow on frontpage
-- [x] Add projects and all gifs to it, plus descriptions and github repos
-- [x] Add shortened About section
-- [ ] Update About me text
-- [ ] Rework Projects page
-- [x] Make John P. Correia on Navbar a clickable button
-
-
-
-# ~~Home~~
-### ~~Set "John P Correia on the top left of navbar to lead back to Home~~
-### ~~Display GIFs of projects [games, webapps, PM projects]~~
-### ~~honestly I would have those gifs of your game as close to front and center as possible. Maybe on the home page a short about section~~
-### ~~“My name is John and I’m a software engineer hobbyist turning professional. I love building games (C#, Unity), but can build about anything (JS, React, Python, C, Node, Express, etc.)”~~
-### ~~Remove button in nav and move everything to Contact section~~
-
-# About Me
-### Change descriptions to talk more about being a developer than being in customer service
-```
-I would work on your “about” quite a bit. 
-It’s not punchy enough and 
-I’m not sure how relevant it is that you’ve been in customer service and theme parks for eight years, 
-or that you enjoyed the experience.
-
-You want something that really says 
-  “I was destined to be a web developer”
-```
-
-# Projects
-### Think about a different card/container for games to show off website as gifs
-### Should say "Q and W to rotate"
-### Display GIFs of projects [games, webapps, PM projects]
-### Add more gifs of my projects
-```
-On the cards for your projects, 
-this is a nice-to-have, 
-but I would include a photo of the project itself 
-maybe choose a different card type for the material design, 
-mostly because your games are beautiful
-
-oh my gosh @John C. this gif on your “hobbies” is AMAZING
-Create one of those for all of your games and put it in the projects section!!!!!!!
-
-small thing - on your tetris game “z” to rotate doesn’t work
-```
-
