+++ package.json
@@ -1,24 +1,19 @@
 {
-  "name": "portfolio",
-  "version": "1.0.0",
-  "description": "Portfolio website",
+  "name": "portfolio-v2",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.6.3",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-scripts": "4.0.2",
-    "web-vitals": "^1.1.0"
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "4.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "build-v2": "cs porfolio-v2 && npm run build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "main": "src/index.js",
   "eslintConfig": {
     "extends": [
       "react-app",
@@ -36,8 +31,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "repository": "https://github.com/JohnJohnx4/portfolio.git",
-  "author": "John Correia <johnjohnx4@gmail.com>",
-  "license": "MIT"
+  }
 }
