--- src/components/About/ImageTiles.js
@@ -1,45 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import GridList from '@material-ui/core/GridList';
-import GridListTile from '@material-ui/core/GridListTile';
-import {tileData} from './tileData';
-
-const styles = theme => ({
-  root: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    justifyContent: 'space-around',
-    overflow: 'hidden',
-    backgroundColor: theme.palette.background.paper,
-  },
-  gridList: {
-    width: 800,
-    height: 410,
-  },
-  subheader: {
-    width: '100%',
-  },
-});
-
-function ImageTiles(props) {
-  const { classes } = props;
-
-  return (
-    <div className={classes.root}>
-      <GridList cellHeight={200} className={classes.gridList} cols={4}>
-        {tileData.map(tile => (
-          <GridListTile key={tile.img} cols={tile.cols || 1}>
-            <img src={tile.img} alt={tile.title} />
-          </GridListTile>
-        ))}
-      </GridList>
-    </div>
-  );
-}
-
-ImageTiles.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(styles)(ImageTiles);
