--- src/components/About/tileData.js
@@ -1,49 +1,0 @@
-export const tileData = [
-	{
-		img: "https://scontent.ftpa1-2.fna.fbcdn.net/v/t1.0-9/29468071_10160196731035397_8786963351960813568_n.jpg?_nc_cat=0&oh=86a96bc2aa7846ed1885d713d4a04063&oe=5BD2FCC0",
-		title: "Image",
-		author: "author",
-		cols: 1
-	},
-	{
-		img: "https://scontent.ftpa1-2.fna.fbcdn.net/v/t1.0-9/13133122_1258555557507674_7984969684232086703_n.jpg?_nc_cat=0&oh=8a586e290b61f89276a0bbeac8d08e4c&oe=5BE3EF14",
-		title: "image",
-		cols: 1
-	},
-	{
-		img: "https://scontent.ftpa1-2.fna.fbcdn.net/v/t1.0-9/10502447_10155024507360397_632960911058575102_n.jpg?_nc_cat=0&oh=1cc12f44252c6663395f3e6a5ece6adb&oe=5BA26122",
-		title: "Image",
-		author: "author",
-		cols: 1
-	},
-	{
-		img: "https://scontent.ftpa1-2.fna.fbcdn.net/v/t1.0-9/12729123_936938783088337_1631797440218329697_n.jpg?_nc_cat=0&oh=82b96887ea954594a6957f0f2948790f&oe=5BD36B6F",
-		title: "Image",
-		author: "author",
-		cols: 1
-	},
-	{
-		img: "https://scontent.ftpa1-2.fna.fbcdn.net/v/t1.0-9/33743044_2083135251716363_3416399918485471232_n.jpg?_nc_cat=0&oh=3a8a06fa1a0a2db316817dad6a39bf7c&oe=5BD57122",
-		title: "Image",
-		author: "author",
-		cols: 1
-	},
-	{
-		img: "https://scontent.ftpa1-2.fna.fbcdn.net/v/t1.0-9/22196465_10159453629560397_8788438839935493270_n.jpg?_nc_cat=0&oh=da01a8a3984a04e80a55e7db65ddd7d8&oe=5BDF22BF",
-		title: "Image",
-		author: "author",
-		cols: 1
-	},
-	{
-		img: "https://scontent.ftpa1-2.fna.fbcdn.net/v/t31.0-8/12087053_10156214569020397_9162444892834957104_o.jpg?_nc_cat=0&oh=6eae8a02f50e28d91fd38f04553bbc3c&oe=5BE614D8",
-		title: "Image",
-		author: "author",
-		cols: 1
-	},
-	{
-		img: "https://scontent.ftpa1-2.fna.fbcdn.net/v/t31.0-8/25182311_10155213658581452_6902965228719617146_o.jpg?_nc_cat=0&oh=873e902cb18dfc1e63eb127d147d6733&oe=5BEBFBFD",
-		title: "Image",
-		author: "author",
-		cols: 1
-	},
-];
