--- src/components/Contact/ContactPage.js
@@ -1,14 +1,0 @@
-import React, { Component } from "react";
-import Contact from "./Contact";
-
-class ContactPage extends Component {
-	render() {
-		return (
-			<div className="contact">
-				<Contact />
-			</div>
-		);
-	}
-}
-
-export default ContactPage;
