--- src/components/Home/HomePage.js
@@ -1,14 +1,0 @@
-import React, { Component } from "react";
-import Home from "./Home";
-
-class HomePage extends Component {
-	render() {
-		return (
-			<div className="homepage">
-				<Home />
-			</div>
-		);
-	}
-}
-
-export default HomePage;
