--- src/components/Navigation/Navigation.js
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-import ApplicationBar from './ApplicationBar';
-
-
-class Navigation extends Component {
-  render() {
-    return (
-      <div className="navbar">
-      <ApplicationBar />
-      </div>
-    );
-  }
-}
-
-export default Navigation;
-
