--- src/components/Projects/ProjectsPage.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-import Projects from './Projects';
-
-class ProjectsPage extends Component {
-  render() {
-    return (
-      <div className="projects">
-        <Projects />
-      </div>
-    );
-  }
-}
-
-export default ProjectsPage;
