--- src/styles/css/index.css
@@ -1,428 +1,0 @@
-/* http://meyerweb.com/eric/tools/css/reset/ 
-   v2.0 | 20110126
-   License: none (public domain)
-*/
-@import url('https://fonts.googleapis.com/css?family=Roboto');
-html,
-body,
-div,
-span,
-applet,
-object,
-iframe,
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-p,
-blockquote,
-pre,
-a,
-abbr,
-acronym,
-address,
-big,
-cite,
-code,
-del,
-dfn,
-em,
-img,
-ins,
-kbd,
-q,
-s,
-samp,
-small,
-strike,
-strong,
-sub,
-sup,
-tt,
-var,
-b,
-u,
-i,
-center,
-dl,
-dt,
-dd,
-ol,
-ul,
-li,
-fieldset,
-form,
-label,
-legend,
-table,
-caption,
-tbody,
-tfoot,
-thead,
-tr,
-th,
-td,
-article,
-aside,
-canvas,
-details,
-embed,
-figure,
-figcaption,
-footer,
-header,
-hgroup,
-menu,
-nav,
-output,
-ruby,
-section,
-summary,
-time,
-mark,
-audio,
-video {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  font-size: 100%;
-  font: inherit;
-  vertical-align: baseline;
-}
-/* HTML5 display-role reset for older browsers */
-article,
-aside,
-details,
-figcaption,
-figure,
-footer,
-header,
-hgroup,
-menu,
-nav,
-section {
-  display: block;
-}
-body {
-  line-height: 1;
-}
-ol,
-ul {
-  list-style: none;
-}
-blockquote,
-q {
-  quotes: none;
-}
-blockquote:before,
-blockquote:after,
-q:before,
-q:after {
-  content: '';
-  content: none;
-}
-table {
-  border-collapse: collapse;
-  border-spacing: 0;
-}
-* {
-  box-sizing: border-box;
-}
-a:link,
-a:visited,
-a:hover,
-a:active {
-  text-decoration: none;
-  color: inherit;
-}
-.homepage {
-  margin: auto;
-}
-.homepage .home .home__grid__container {
-  width: 100%;
-  margin: inherit;
-}
-.homepage .home .home__grid__container .home__grid {
-  padding: 0;
-}
-.homepage .home .home__grid__container .home__grid .home__paper .home__avatar {
-  width: 100px;
-  height: 100px;
-}
-.homepage .home .home__grid__container .home__grid .home__paper .home__avatar img {
-  border-radius: 50%;
-  width: 100px;
-  height: 100px;
-}
-@media (max-width: 479px) {
-  .homepage {
-    max-width: 100%;
-  }
-  .homepage .home {
-    width: 100%;
-  }
-  .homepage .home .home__grid__container {
-    width: 90%;
-    margin: auto;
-  }
-  .homepage .home .home__grid__container .home__grid {
-    padding: 0;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar {
-    width: 50px;
-    height: 50px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar img {
-    border-radius: 50%;
-    width: 50px;
-    height: 50px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__header {
-    font-size: 14px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__subheader {
-    font-size: 12px;
-  }
-}
-@media (min-width: 480px) {
-  .homepage {
-    max-width: 100%;
-  }
-  .homepage .home .home__grid__container {
-    width: 80%;
-    margin: auto;
-  }
-  .homepage .home .home__grid__container .home__grid {
-    padding: 0;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar {
-    width: 75px;
-    height: 75px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar img {
-    border-radius: 50%;
-    width: 75px;
-    height: 75px;
-  }
-}
-@media (min-width: 900px) {
-  .homepage {
-    max-width: 70%;
-  }
-  .homepage .home .home__grid__container {
-    width: 100%;
-    margin: auto;
-  }
-  .homepage .home .home__grid__container .home__grid {
-    padding: 0;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar {
-    width: 100px;
-    height: 100px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar img {
-    border-radius: 50%;
-    width: 100px;
-    height: 100px;
-  }
-}
-@media (min-width: 1600px) {
-  .homepage {
-    max-width: 65%;
-  }
-  .homepage .home .home__grid__container {
-    width: 100%;
-    margin: auto;
-  }
-  .homepage .home .home__grid__container .home__grid {
-    padding: 0;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar {
-    width: 125px;
-    height: 125px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar img {
-    border-radius: 50%;
-    width: 125px;
-    height: 125px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__header {
-    font-size: 24px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__subheader {
-    font-size: 22px;
-  }
-}
-@media (min-width: 2560px) {
-  .homepage {
-    max-width: 60%;
-  }
-  .homepage .home .home__grid__container {
-    width: 100%;
-    margin: inherit;
-  }
-  .homepage .home .home__grid__container .home__grid {
-    padding: 0;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper {
-    padding: 5%;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar {
-    width: 150px;
-    height: 150px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__avatar img {
-    border-radius: 50%;
-    width: 150px;
-    height: 150px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__header {
-    font-size: 44px;
-  }
-  .homepage .home .home__grid__container .home__grid .home__paper .home__subheader {
-    font-size: 32px;
-  }
-}
-.experience__container {
-  padding: 10px !important;
-}
-@media (max-width: 479px) {
-  .experienece {
-    margin: 0 10px !important;
-  }
-  .experience__header {
-    font-size: 12px !important;
-  }
-  .experience__subheader {
-    font-size: 10px !important;
-  }
-}
-@media (max-width: 479px) {
-  .about__container {
-    margin: 0 10px !important;
-  }
-}
-@media (max-width: 479px) {
-  .about__grid {
-    display: none !important;
-  }
-}
-.about__grid2 {
-  display: none !important;
-}
-@media (max-width: 479px) {
-  .about__grid2 {
-    display: block !important;
-  }
-  .about__grid2 .about__text__grid {
-    height: auto;
-  }
-  .about__grid2 .about__text__grid .about__text {
-    height: 100%;
-  }
-}
-.about__img__grid .about__img__paper .about__image {
-  max-width: 200px;
-  max-height: 200px;
-  margin: auto;
-  border: 0;
-  border-radius: 50%;
-}
-@media (max-width: 479px) {
-  .about__img__grid {
-    max-height: 20%;
-  }
-  .about__img__grid .about__img__paper {
-    height: 120px;
-  }
-  .about__img__grid .about__img__paper .about__image {
-    max-height: 90%;
-    margin: none;
-    border-radius: 50%;
-  }
-  .about__img__grid .about__img__paper .about__image img {
-    margin: 0;
-  }
-}
-@media (max-width: 479px) {
-  .friends__container {
-    flex-wrap: wrap;
-  }
-}
-.navbar {
-  margin-bottom: 10px;
-}
-@media (max-width: 479px) {
-  .navbar {
-    margin-bottom: 5px;
-    min-width: 100%;
-    max-width: 100%;
-  }
-  .navbar .appbar .appbar__container {
-    width: 100%;
-  }
-  .navbar .appbar .appbar__container .appbar__title {
-    font-size: 16px;
-  }
-  .navbar .appbar .appbar__container .appbar__button {
-    font-size: 9px;
-    margin: 5px;
-    min-width: 25px;
-  }
-  .navbar button {
-    font-size: 10px;
-    padding: 0;
-  }
-}
-.projects {
-  max-width: 900px;
-  margin: auto;
-}
-@media (max-width: 479px) {
-  .projects {
-    max-width: 90%;
-    min-width: 110px;
-  }
-  .projects .project__icon {
-    margin-right: 5px;
-    width: 20px;
-    height: 20px;
-  }
-  .projects .project__title {
-    padding-left: 10px;
-    min-width: 100px;
-  }
-  .projects .project__title span {
-    font-size: 14px;
-  }
-  .projects .project__text {
-    font-size: 10px;
-  }
-}
-.contact {
-  max-width: 900px;
-  margin: auto;
-}
-@media (max-width: 479px) {
-  .contact {
-    max-width: 90%;
-    min-width: 110px;
-  }
-  .contact .contact__icon {
-    margin: 0;
-  }
-  .contact .contact__title span {
-    font-size: 13px;
-    font-weight: bold;
-  }
-  .contact .contact__text {
-    font-size: 10px;
-  }
-}
-body {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-}
