--- src/styles/less/App.less
@@ -1,3 +1,0 @@
-.App {
-
-}
