--- src/styles/less/about.less
@@ -1,64 +1,0 @@
-@smartphones: ~"(max-width: 479px)";
-@tablet: ~"(min-width: 480px)";
-@laptop: ~"(min-width: 900px)";
-@desktop: ~"(min-width: 1600px)";
-@monster: ~"(min-width: 2560px)";
-
-
-.about__container {
-  @media @smartphones {
-  margin: 0 10px !important;
-}}
-
-.about__grid {
-  @media @smartphones {
-    display: none !important;
-  }
-}
-
-.about__grid2 {
-  display: none !important;
-  @media @smartphones {
-    display: block !important;
-    .about__text__grid {
-      height: auto;
-      .about__text {
-        height: 100%;
-      }
-    }
-  }
-}
-
-.about__img__grid {
-  .about__img__paper {
-    .about__image {
-      max-width: 200px;
-      max-height: 200px;
-      margin: auto;
-      border: 0;
-      border-radius: 50%;
-    }
-  }
-  @media @smartphones {
-  // about__img__grid
-    max-height: 20%;
-    .about__img__paper {
-      height: 120px;
-      .about__image {
-        max-height: 90%;
-        margin: none;
-        border-radius: 50%;
-        img {
-          margin: 0;
-        }
-      }
-    }
-  }
-}
-
-
-@media @smartphones {
-  .friends__container{
-    flex-wrap: wrap;
-  }
-}
