--- src/styles/less/contact.less
@@ -1,36 +1,0 @@
-@smartphones: ~"(max-width: 479px)";
-@tablet: ~"(min-width: 480px)";
-@laptop: ~"(min-width: 900px)";
-@desktop: ~"(min-width: 1600px)";
-@monster: ~"(min-width: 2560px)";
-
-.contact {
-    max-width: 900px;
-    margin: auto;
-    @media @smartphones {
-      max-width: 90%;
-      min-width: 110px;
-      .contact__icon {
-        margin: 0;
-      }
-      .contact__title {
-        span {
-          font-size: 13px;
-          font-weight: bold;
-        }
-      }
-      .contact__text {
-        font-size: 10px;
-      }
-      
-    }
-  .contact__icon {
-
-  }
-  .contact__title {
-
-  }
-  .contact__text {
-
-  }
-  }
