--- src/styles/less/experience.less
@@ -1,23 +1,0 @@
-@smartphones: ~"(max-width: 479px)";
-@tablet: ~"(min-width: 480px)";
-@laptop: ~"(min-width: 900px)";
-@desktop: ~"(min-width: 1600px)";
-@monster: ~"(min-width: 2560px)";
-
-.experience__container {
-    padding: 10px !important;
-}
-
-@media @smartphones {
-  .experienece {
-    margin: 0 10px !important;
-  }
-
-  .experience__header {
-    font-size: 12px !important;
-  }
-
-  .experience__subheader {
-    font-size: 10px !important;
-  }
-}
