--- src/styles/less/home.less
@@ -1,185 +1,0 @@
-@smartphones: ~"(max-width: 479px)";
-@tablet: ~"(min-width: 480px)";
-@laptop: ~"(min-width: 900px)";
-@desktop: ~"(min-width: 1600px)";
-@monster: ~"(min-width: 2560px)";
-
-a:link,
-a:visited,
-a:hover,
-a:active {
-	text-decoration: none;
-	color: inherit;
-}
-
-.homepage {
-	margin: auto;
-	.home {
-		.home__grid__container {
-			width: 100%;
-			margin: inherit;
-			.home__grid {
-				padding: 0;
-				.home__paper {
-					.home__avatar {
-						width: 100px;
-						height: 100px;
-						img {
-							border-radius: 50%;
-							width: 100px;
-							height: 100px;
-						}
-					}
-					.home__header {
-					}
-					.home__subheader {
-					}
-				}
-			}
-		}
-	}
-
-	@media @smartphones {
-		max-width: 100%;
-		.home {
-			width: 100%;
-			.home__grid__container {
-				width: 90%;
-				margin: auto;
-				.home__grid {
-					padding: 0;
-					.home__paper {
-						.home__avatar {
-							width: 50px;
-							height: 50px;
-							img {
-								border-radius: 50%;
-								width: 50px;
-								height: 50px;
-							}
-						}
-						.home__header {
-							font-size: 14px;
-						}
-						.home__subheader {
-							font-size: 12px;
-						}
-					}
-				}
-			}
-		}
-	}
-	@media @tablet {
-		max-width: 100%;
-		.home {
-			.home__grid__container {
-				width: 80%;
-				margin: auto;
-				.home__grid {
-					padding: 0;
-					.home__paper {
-						.home__avatar {
-							width: 75px;
-							height: 75px;
-							img {
-								border-radius: 50%;
-								width: 75px;
-								height: 75px;
-							}
-						}
-						.home__header {
-						}
-						.home__subheader {
-						}
-					}
-				}
-			}
-		}
-	}
-	@media @laptop {
-		max-width: 70%;
-		.home {
-			.home__grid__container {
-				width: 100%;
-				margin: auto;
-				.home__grid {
-					padding: 0;
-					.home__paper {
-						.home__avatar {
-							width: 100px;
-							height: 100px;
-							img {
-								border-radius: 50%;
-								width: 100px;
-								height: 100px;
-							}
-						}
-						.home__header {
-						}
-						.home__subheader {
-						}
-					}
-				}
-			}
-		}
-	}
-	@media @desktop {
-		max-width: 65%;
-		.home {
-			.home__grid__container {
-				width: 100%;
-				margin: auto;
-				.home__grid {
-					padding: 0;
-					.home__paper {
-						.home__avatar {
-							width: 125px;
-							height: 125px;
-							img {
-								border-radius: 50%;
-								width: 125px;
-								height: 125px;
-							}
-						}
-						.home__header {
-							font-size: 24px;
-						}
-						.home__subheader {
-							font-size: 22px;
-						}
-					}
-				}
-			}
-		}
-	}
-	@media @monster {
-		max-width: 60%;
-		.home {
-			.home__grid__container {
-				width: 100%;
-				margin: inherit;
-				.home__grid {
-					padding: 0;
-					.home__paper {
-						padding: 5%;
-						.home__avatar {
-							width: 150px;
-							height: 150px;
-							img {
-								border-radius: 50%;
-								width: 150px;
-								height: 150px;
-							}
-						}
-						.home__header {
-							font-size: 44px;
-						}
-						.home__subheader {
-							font-size: 32px;
-						}
-					}
-				}
-			}
-		}
-	}
-}
