--- src/styles/less/index.less
@@ -1,16 +1,0 @@
-@import 'App.less';
-@import 'reset.less';
-@import 'home.less';
-@import 'experience.less';
-@import 'about.less';
-@import 'navigation.less';
-@import 'projects.less';
-@import 'contact.less';
-@import url('https://fonts.googleapis.com/css?family=Roboto');
-
-body {
-    margin: 0;
-    padding: 0;
-    font-family: sans-serif;
-  }
-
