--- src/styles/less/navigation.less
@@ -1,33 +1,0 @@
-@smartphones: ~"(max-width: 479px)";
-@tablet: ~"(min-width: 480px)";
-@laptop: ~"(min-width: 900px)";
-@desktop: ~"(min-width: 1600px)";
-@monster: ~"(min-width: 2560px)";
-
-.navbar {
-	margin-bottom: 50px; 
-	
-	@media @smartphones {
-		margin-bottom: 25px; 
-		min-width: 100%;
-		max-width: 100%;
-		.appbar {
-			.appbar__container {
-				width:100%;
-				.appbar__title {
-					font-size: 16px;
-				}
-
-				.appbar__button {
-					font-size: 9px;
-					margin: 5px;
-					min-width: 25px;
-				}
-			}
-		}
-		button {
-			font-size: 10px;
-			padding: 0;
-		}
-	}
-}
