--- src/styles/less/projects.less
@@ -1,31 +1,0 @@
-@smartphones: ~"(max-width: 479px)";
-@tablet: ~"(min-width: 480px)";
-@laptop: ~"(min-width: 900px)";
-@desktop: ~"(min-width: 1600px)";
-@monster: ~"(min-width: 2560px)";
-
-.projects {
-    max-width: 900px;
-    margin: auto;
-    @media @smartphones {
-      max-width: 90%;
-      min-width: 110px;
-      
-      .project__icon{
-        margin-right: 5px;
-        width: 20px;
-        height: 20px;
-      }
-      .project__title{
-        padding-left: 10px;
-        min-width: 100px;
-        span {
-          font-size: 14px;
-        }
-      }
-      .project__text{
-        font-size: 10px;
-        
-      }
-    }
-  }
